+++ plugin-dev/Source/Sentry/Private/Apple/AppleSentrySubsystem.cpp
@@ -12,9 +12,11 @@
 #include "SentrySamplingContextApple.h"
 #include "SentryIdApple.h"
 
+#include "SentryBreadcrumb.h"
 #include "SentryEvent.h"
 #include "SentrySettings.h"
 #include "SentryBeforeSendHandler.h"
+#include "SentryBeforeBreadcrumbHandler.h"
 #include "SentryDefines.h"
 #include "SentrySamplingContext.h"
 #include "SentryTraceSampler.h"
@@ -34,7 +36,7 @@
 #include "UObject/UObjectThreadContext.h"
 #include "Utils/SentryLogUtils.h"
 
-void FAppleSentrySubsystem::InitWithSettings(const USentrySettings* settings, USentryBeforeSendHandler* beforeSendHandler, USentryTraceSampler* traceSampler)
+void FAppleSentrySubsystem::InitWithSettings(const USentrySettings* settings, USentryBeforeSendHandler* beforeSendHandler, USentryBeforeBreadcrumbHandler* beforeBreadcrumbHandler, USentryTraceSampler* traceSampler)
 {
 	[SENTRY_APPLE_CLASS(PrivateSentrySDKOnly) setSdkName:@"sentry.cocoa.unreal"];
 
@@ -43,6 +45,11 @@ void FAppleSentrySubsystem::InitWithSettings(const USentrySettings* settings, US
 	dispatch_async(dispatch_get_main_queue(), ^{
 		[SENTRY_APPLE_CLASS(SentrySDK) startWithConfigureOptions:^(SentryOptions *options) {
 			options.dsn = settings->Dsn.GetNSString();
+#if WITH_EDITOR
+			if(!settings->EditorDsn.IsEmpty()) {
+				options.dsn = settings->EditorDsn.GetNSString();
+			}
+#endif
 			options.environment = settings->Environment.GetNSString();
 			options.enableAutoSessionTracking = settings->EnableAutoSessionTracking;
 			options.sessionTrackingIntervalMillis = settings->SessionTimeout;
@@ -76,11 +83,18 @@ void FAppleSentrySubsystem::InitWithSettings(const USentrySettings* settings, US
 				}];
 			};
 			options.beforeSend = ^SentryEvent* (SentryEvent* event) {
-				FGCScopeGuard GCScopeGuard;
-
 				if (FUObjectThreadContext::Get().IsRoutingPostLoad)
 				{
 					UE_LOG(LogSentrySdk, Log, TEXT("Executing `beforeSend` handler is not allowed when post-loading."));
+					return event;
+				}
+
+				if (IsGarbageCollecting())
+				{
+					// If event is captured during garbage collection we can't instantiate UObjects safely or obtain a GC lock
+					// since it will cause a deadlock (see https://github.com/getsentry/sentry-unreal/issues/850).
+					// In this case event will be reported without calling a `beforeSend` handler.
+					UE_LOG(LogSentrySdk, Log, TEXT("Executing `beforeSend` handler is not allowed during garbage collection."));
 					return event;
 				}
 
@@ -110,6 +124,30 @@ void FAppleSentrySubsystem::InitWithSettings(const USentrySettings* settings, US
 					USentrySamplingContext* Context = USentrySamplingContext::Create(MakeShareable(new SentrySamplingContextApple(samplingContext)));
 					float samplingValue;
 					return traceSampler->Sample(Context, samplingValue) ? [NSNumber numberWithFloat:samplingValue] : nil;
+				};
+			}
+			if (beforeBreadcrumbHandler != nullptr)
+			{
+				options.beforeBreadcrumb = ^SentryBreadcrumb* (SentryBreadcrumb* breadcrumb) {
+					if (FUObjectThreadContext::Get().IsRoutingPostLoad)
+					{
+						// Don't print to logs within `onBeforeBreadcrumb` handler as this can lead to creating new breadcrumb
+						return breadcrumb;
+					}
+
+					if (IsGarbageCollecting())
+					{
+						// If breadcrumb is added during garbage collection we can't instantiate UObjects safely or obtain a GC lock
+						// since there is no guarantee it will be ever freed.
+						// In this case breadcrumb will be added without calling a `beforeBreadcrumb` handler.
+						return breadcrumb;
+					}
+
+					USentryBreadcrumb* BreadcrumbToProcess = USentryBreadcrumb::Create(MakeShareable(new SentryBreadcrumbApple(breadcrumb)));
+
+					USentryBreadcrumb* ProcessedBreadcrumb = beforeBreadcrumbHandler->HandleBeforeBreadcrumb(BreadcrumbToProcess, nullptr);
+
+					return ProcessedBreadcrumb ? breadcrumb : nullptr;
 				};
 			}
 		}];